refactor: extract event topology bootverify checks
